+++ NEWS.md
@@ -3,6 +3,7 @@
 - Fix bug in `infer_data_types()` to avoid errors with factors/numeric values
 - Improvements to documentation, error checking
 - Allow user to access column mapping for questions and IDs (#115)
+- Deprecate `registerOptions()` to avoid unexpected behavior with options
 
 # qualtRics 3.1.0